General Structure
Just the basic idea of the components used. One component for each input, (Since it's going to have some dynamic restrictions on input with JS) one component for each row of input (to add more dynamically when they're all filled) and one component for the form overall (potentially to allow tracking multiple casettes).