Feat/structural setup (#1)
* General Structure

Just the basic idea of the components used. One component for each input, (Since it's going to have some dynamic restrictions on input with JS) one component for each row of input (to add more dynamically when they're all filled) and one component for the form overall (potentially to allow tracking multiple casettes).

* Form of Input Handling

No actual intended effects here, but the function layout I've made here should work once we get underway

* Remaining Components

The remaining structure of the website should be in place now, so we're ready for dev, as far as I can tell.