Dynamicism and Capsulization
Made it so the user can add more rows with the add row button, and made it so all the components contain is their actual html, encapsulating all the scripts in a folder for greater modularity. This is the last piece of input management, so I'll be merging to dev now.