Feat/input management (#2)
* Initial Pass

This has the first part of the intended functionality, only allowing numbers (though it's a bit fiddly). Tomorrow, I'll start only allowing 0-60 as input values.

* Giving up, sorta

Since html has a number input that does what I was initially trying to do, I'm instead going to come at this from a different angle, but here's a "functioning" increment.

* Seconds Function

The seconds functions as intended now! There needs to be more to it, to ensure compatability with things that don't support the number input type, but it's a good starting point.

* Fallback Code

Some browsers don't support the number type, and so it'd fall back to text input. This code allows the input box to sanitize the input, so we only get values from 0 to 59. (If they input 60, it just increments the minute)

* Minutes Fallback

The minutes now has fallback code that only allows digits, so that input is sanitized as well. Potentially come back to minutes to put a max that's equal to the selected casette length?

* Adding rows, sorta

The add row button does, in fact, add a row... one time, then never works again. As far as I can tell, the method I'm using isn't duplicating the scripts I've tied to each element. Will work on rebinding those after lunch (might require making a js file that has all my scripts that I pull from, if that's possible that'd be sick)

* Dynamicism and Capsulization

Made it so the user can add more rows with the add row button, and made it so all the components contain is their actual html, encapsulating all the scripts in a folder for greater modularity. This is the last piece of input management, so I'll be merging to dev now.