Feat/time calculation (#3)
* Intended Functionality

Adds/changes a text node at the bottom of the page that tells you the total time. This should be all that's required of this branch, so we're able to merge into dev and work on the last initially required feature: the casette length setting.

* Minor Visual Change

The display would just display single digit seconds a-la :2 instead of the more-sensible :02, fixed that. For real, ready for dev now. Thank goodness for the magic of regex.