Halfway Marker + Limit Sensor
Adds a marker at the end of side a, and changes the bottom text's class if you've overrun the time available on your casette. If you don't have a length selected, it doesn't do any of this. Should be ready for dev now. If I'm feeling ambitious monday, I'll add fancy css touches.