Initial media queries
It's not quite done yet, but the piano keys are done! Note the casette numbering and the piano key labelling needs work, but the keys themselves are sized properly. (The other two interactable elements need some moving as well. Let's hope you can change keyframes based on media queries in the case of the new row button!)