Release 1.0.0 (#7)
* Feat/structural setup (#1)

** General Structure

Just the basic idea of the components used. One component for each input, (Since it's going to have some dynamic restrictions on input with JS) one component for each row of input (to add more dynamically when they're all filled) and one component for the form overall (potentially to allow tracking multiple casettes).

** Form of Input Handling

No actual intended effects here, but the function layout I've made here should work once we get underway

** Remaining Components

The remaining structure of the website should be in place now, so we're ready for dev, as far as I can tell.

* Feat/input management (#2)

** Initial Pass

This has the first part of the intended functionality, only allowing numbers (though it's a bit fiddly). Tomorrow, I'll start only allowing 0-60 as input values.

** Giving up, sorta

Since html has a number input that does what I was initially trying to do, I'm instead going to come at this from a different angle, but here's a "functioning" increment.

** Seconds Function

The seconds functions as intended now! There needs to be more to it, to ensure compatability with things that don't support the number input type, but it's a good starting point.

** Fallback Code

Some browsers don't support the number type, and so it'd fall back to text input. This code allows the input box to sanitize the input, so we only get values from 0 to 59. (If they input 60, it just increments the minute)

** Minutes Fallback

The minutes now has fallback code that only allows digits, so that input is sanitized as well. Potentially come back to minutes to put a max that's equal to the selected casette length?

** Adding rows, sorta

The add row button does, in fact, add a row... one time, then never works again. As far as I can tell, the method I'm using isn't duplicating the scripts I've tied to each element. Will work on rebinding those after lunch (might require making a js file that has all my scripts that I pull from, if that's possible that'd be sick)

** Dynamicism and Capsulization

Made it so the user can add more rows with the add row button, and made it so all the components contain is their actual html, encapsulating all the scripts in a folder for greater modularity. This is the last piece of input management, so I'll be merging to dev now.

* Feat/time calculation (#3)

** Intended Functionality

Adds/changes a text node at the bottom of the page that tells you the total time. This should be all that's required of this branch, so we're able to merge into dev and work on the last initially required feature: the casette length setting.

** Minor Visual Change

The display would just display single digit seconds a-la :2 instead of the more-sensible :02, fixed that. For real, ready for dev now. Thank goodness for the magic of regex.

* Halfway Marker + Limit Sensor (#4)

Adds a marker at the end of side a, and changes the bottom text's class if you've overrun the time available on your casette. If you don't have a length selected, it doesn't do any of this. Should be ready for dev now. If I'm feeling ambitious monday, I'll add fancy css touches.

* Feat/css fanciness (#5)

** Initial Commit

Reorganizing some things to make it a little cleaner on my end, and detailing the basic concepts I want to go for. Ideally the general look is going to be of an old hi-fi section.

** Basic ideas

General color setup, will work on alignment and better looks after lunch

** Animations

When you add a new row, and when you hit the calculate button, there's a real neat "unfolding component" animation now. Up next is the piano key styling on the radio buttons. (Gonna take some heavy overriding)

** Casette Length Styling

After a great deal of fiddling with the css for radio buttons, I've got nice fancy piano key looking buttons. I even used the linear-gradient function to make them look "pressed in".  Up next is the woodgrain to replace that drab wood color, font choice, and mobile responsiveness. Mobile responsiveness should only require overriding some margins/padding, everything else should fit, regardless of display width.

** Final Background Changes

Added some fonts and woodgrain! Now all that's left is mobile responsiveness, but that's a problem for another branch and another day.

* Feat/responsiveness and bugs (#6)

** Bug Fixes

Fixed some bugs in the actual function of the website. Now comes the hard part: media queries.

** Initial media queries

It's not quite done yet, but the piano keys are done! Note the casette numbering and the piano key labelling needs work, but the keys themselves are sized properly. (The other two interactable elements need some moving as well. Let's hope you can change keyframes based on media queries in the case of the new row button!)

** Additional Media Queries

So I had to fix the way the piano keys are labelled, and for some reason the 110 key's label doesn't display correctly. Unfortunately, that means I had to use absolute positioning, but hey, I calibrated the one at 550 px width on my actual cell phone, so there's that! Additionally I removed the mediaqueries.css file, since I couldn't keep them all in a separate file and still have them work as intended (so I crammed 'em all in the lengthpicker, confusingly for the things that aren't lengthpicker keys). Should be ready for dev, then dev ready for master now! yay!